scripts/generate_supplementary.py
"""
Generate Supplementary Materials DOCX with embedded figures
"""

import pandas as pd
from docx import Document
from docx.shared import Pt, Inches
from docx.enum.text import WD_ALIGN_PARAGRAPH
from docx.oxml.ns import qn
from docx.oxml import OxmlElement
from pathlib import Path

BASE_DIR = Path(__file__).parent.parent

def set_cell_shading(cell, color):
    tc = cell._tc
    tcPr = tc.get_or_add_tcPr()
    shd = OxmlElement('w:shd')
    shd.set(qn('w:fill'), color)
    tcPr.append(shd)

def create_supplementary():
    doc = Document()
    
    style = doc.styles['Normal']
    style.font.name = 'Times New Roman'
    style.font.size = Pt(11)
    
    # Title
    title = doc.add_heading('', level=0)
    run = title.add_run('Supplementary Materials')
    run.font.size = Pt(16)
    run.bold = True
    title.alignment = WD_ALIGN_PARAGRAPH.CENTER
    
    doc.add_paragraph()
    
    subtitle = doc.add_paragraph()
    subtitle.add_run('Phase-Specific Host-Directed Therapy Targets in Sepsis: An Integrated Multi-omics and Chemoinformatics Pipeline').italic = True
    subtitle.alignment = WD_ALIGN_PARAGRAPH.CENTER
    
    doc.add_paragraph()
    doc.add_paragraph('Siddalingaiah H S')
    
    doc.add_page_break()
    
    # Table of Contents
    doc.add_heading('Table of Contents', level=1)
    doc.add_paragraph('Supplementary Table S1: Complete 60-Gene Sepsis Host Signature')
    doc.add_paragraph('Supplementary Table S2: All 37 Drug Candidates with Clinical Stage')
    doc.add_paragraph('Supplementary Table S3: Literature Validation Summary')
    doc.add_paragraph('Supplementary Figure S1: Pathway Distribution of Targets')
    doc.add_paragraph('Supplementary Figure S2: Compound Potency by Target')
    
    doc.add_page_break()
    
    # ----- TABLE S1 -----
    doc.add_heading('Supplementary Table S1: Complete 60-Gene Sepsis Host Signature', level=1)
    
    targets_df = pd.read_csv(BASE_DIR / 'outputs' / 'tables' / 'targets_ranked.csv')
    
    table1 = doc.add_table(rows=len(targets_df)+1, cols=6)
    table1.style = 'Table Grid'
    
    headers = ['Rank', 'Gene', 'Symbol', 'Pathway', 'Phase', 'Score']
    for i, h in enumerate(headers):
        cell = table1.rows[0].cells[i]
        cell.text = h
        cell.paragraphs[0].runs[0].bold = True
        set_cell_shading(cell, 'D9E2F3')
    
    for i, row in targets_df.iterrows():
        table1.rows[i+1].cells[0].text = str(row['Rank'])
        table1.rows[i+1].cells[1].text = row['Gene']
        table1.rows[i+1].cells[2].text = row['Symbol']
        table1.rows[i+1].cells[3].text = row['Pathway']
        table1.rows[i+1].cells[4].text = row['Phase_Relevance']
        table1.rows[i+1].cells[5].text = str(row['Composite_Score'])
    
    doc.add_page_break()
    
    # ----- TABLE S2 -----
    doc.add_heading('Supplementary Table S2: All Drug Candidates with Clinical Development Stage', level=1)
    
    compounds_df = pd.read_csv(BASE_DIR / 'outputs' / 'tables' / 'compounds_ranked.csv')
    
    table2 = doc.add_table(rows=len(compounds_df)+1, cols=5)
    table2.style = 'Table Grid'
    
    headers2 = ['Drug', 'Target', 'pChEMBL', 'Phase', 'Clinical Evidence']
    for i, h in enumerate(headers2):
        cell = table2.rows[0].cells[i]
        cell.text = h
        cell.paragraphs[0].runs[0].bold = True
        set_cell_shading(cell, 'D9E2F3')
    
    for i, row in compounds_df.iterrows():
        table2.rows[i+1].cells[0].text = str(row['Drug'])
        table2.rows[i+1].cells[1].text = str(row['Target'])
        table2.rows[i+1].cells[2].text = str(row['pChEMBL'])
        phase_map = {4: 'Approved', 3: 'Phase III', 2: 'Phase II', 1: 'Phase I'}
        table2.rows[i+1].cells[3].text = phase_map.get(row['Phase'], str(row['Phase']))
        table2.rows[i+1].cells[4].text = str(row['Evidence'])
    
    doc.add_page_break()
    
    # ----- TABLE S3 Summary -----
    doc.add_heading('Supplementary Table S3: Literature Validation Summary', level=1)
    
    p = doc.add_paragraph()
    p.add_run('Search Strategy: ').bold = True
    p.add_run('PubMed search "[Gene Symbol] AND (sepsis OR septic shock)" for each target gene.')
    
    doc.add_paragraph()
    
    validation_summary = [
        ('IL6', '2,847', 'Strong', 'RECOVERY trial success'),
        ('TNF', '3,521', 'Strong', 'Multiple failed trials - phase mismatch'),
        ('TLR4', '1,245', 'Strong', 'ACCESS trial failed'),
        ('PD1', '312', 'Moderate', 'Phase 1b pilot positive'),
        ('NLRP3', '856', 'Strong', 'MCC950 in development'),
        ('IL1B', '2,156', 'Strong', 'Anakinra SAVE-MORE success'),
        ('JAK2', '423', 'Moderate', 'Baricitinib ACTT-2 approved'),
        ('HMGB1', '567', 'Moderate', 'Glycyrrhizin in preclinical'),
        ('STAT3', '389', 'Moderate', 'JAK inhibitors target'),
        ('CASP1', '234', 'Moderate', 'VX-765 in Phase II'),
    ]
    
    table3 = doc.add_table(rows=11, cols=4)
    table3.style = 'Table Grid'
    
    headers3 = ['Gene', 'PubMed Count', 'Validation', 'Key Finding']
    for i, h in enumerate(headers3):
        cell = table3.rows[0].cells[i]
        cell.text = h
        cell.paragraphs[0].runs[0].bold = True
        set_cell_shading(cell, 'D9E2F3')
    
    for i, row_data in enumerate(validation_summary):
        for j, val in enumerate(row_data):
            table3.rows[i+1].cells[j].text = val
    
    doc.add_page_break()
    
    # ----- FIGURES -----
    doc.add_heading('Supplementary Figure S1: Pathway Distribution', level=1)
    doc.add_picture(str(BASE_DIR / 'outputs' / 'figures' / 'figure4_pathway_heatmap.png'), width=Inches(6))
    doc.paragraphs[-1].alignment = WD_ALIGN_PARAGRAPH.CENTER
    
    p = doc.add_paragraph()
    p.add_run('Figure S1: ').bold = True
    p.add_run('(A) Mean composite score by pathway and immune phase. (B) Number of targets per functional pathway.')
    
    doc.add_page_break()
    
    doc.add_heading('Supplementary Figure S2: Compound Potency by Target', level=1)
    doc.add_picture(str(BASE_DIR / 'outputs' / 'figures' / 'figure3_target_potency.png'), width=Inches(6))
    doc.paragraphs[-1].alignment = WD_ALIGN_PARAGRAPH.CENTER
    
    p = doc.add_paragraph()
    p.add_run('Figure S2: ').bold = True
    p.add_run('Maximum compound potency (pChEMBL) for top 15 targets. Vertical lines indicate 1 µM (red) and 10 nM (green) thresholds.')
    
    doc.add_page_break()
    
    doc.add_heading('Supplementary Figure S3: Sepsis Immune Timeline', level=1)
    doc.add_picture(str(BASE_DIR / 'outputs' / 'figures' / 'figure5_sepsis_timeline.png'), width=Inches(6))
    doc.paragraphs[-1].alignment = WD_ALIGN_PARAGRAPH.CENTER
    
    p = doc.add_paragraph()
    p.add_run('Figure S3: ').bold = True
    p.add_run('Schematic representation of biphasic sepsis immune response showing hyperinflammation (0-72h) transitioning to immunosuppression (>72h) with corresponding HDT intervention windows.')
    
    # Save
    output_path = BASE_DIR / 'manuscripts' / 'Supplementary_Materials.docx'
    doc.save(str(output_path))
    print(f'Created: {output_path}')

if __name__ == '__main__':
    create_supplementary()
