pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "sepsis-hdt-pipeline"
version = "1.0.0"
description = "Phase-specific host-directed therapy target identification for sepsis"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Dr. Siddalingaiah H S", email = "hssling@yahoo.com"}
]
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]
dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "python-docx>=0.8.11",
    "requests>=2.28.0",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]

[project.urls]
Homepage = "https://github.com/hssling/Sepsis_drug_discovery"
Documentation = "https://github.com/hssling/Sepsis_drug_discovery/docs"
Repository = "https://github.com/hssling/Sepsis_drug_discovery"

[tool.black]
line-length = 120
target-version = ['py310', 'py311', 'py312']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "-v --cov=scripts"
